+++ src/google/protobuf/stubs/common.cc
@@ -374,7 +374,7 @@ void ShutdownProtobufLibrary() {
   // Make it safe to call this multiple times.
   if (internal::shutdown_functions == NULL) return;
 
-  for (int i = 0; i < internal::shutdown_functions->size(); i++) {
+  for (size_t i = 0; i < internal::shutdown_functions->size(); i++) {
     internal::shutdown_functions->at(i)();
   }
   delete internal::shutdown_functions;
